refactor: 라우팅 계층 중복 제거 및 UI 로직 분리 (Round 16)
- SkillStat, SkillStatConfig, SKILL_STAT_CONFIG, buildSkillStats를 domain/skills.ts에서 lib/skillStats.ts로 이동
- domain/skills.ts: countSkillsByType만 유지
- StatsBar.tsx: 임포트 경로를 @/lib/skillStats로 업데이트
- 테스트 파일 이동 및 정리
