+++ docs/__agents_only__/fix-log.md
@@ -5,11 +5,6 @@
 - Rule: TypeScript — 도메인 유니온 타입이 존재할 때 인터페이스 prop에 `string`을 사용하면 타입 안전성 저하
 - Context: `BacktestCaseCard.tsx`의 내부 `RiskBadgeProps`에서 `level: string` 대신 `level: BacktestRiskLevel`을 사용해야 함
 
-## [PR #344 Round 14 | refactor/343/라우팅-계층-중복-제거-ui-로직-분리 | 2026-04-21]
-- Violation: `domain/types.ts`에 UI 관심사 타입 3종 (`PriceChangeDisplay`, `BreadcrumbItem`, `SkillStat`/`SkillStatConfig`) 정의
-- Rule: ARCHITECTURE.md — domain/ 레이어는 순수 TypeScript 함수만 포함; CSS 클래스명·유니코드 표시문자·접근성 레이블 등 프레젠테이션 관심사를 포함할 수 없음
-- Context: 각 타입의 유일한 소비처(`lib/priceFormat.ts`, `lib/seo.ts`, `domain/skills.ts`)로 이동; domain/types.ts에서 삭제
-
